adjust user-related models
