+++ Studenda.Core/Model/Security/Management/Role.cs
@@ -1,14 +1,12 @@
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Studenda.Core.Data.Configuration;
-using Studenda.Core.Model.Link;
 
-namespace Studenda.Core.Model.Account;
+namespace Studenda.Core.Model.Security.Management;
 
 /// <summary>
-///     Разрешение для <see cref="Role" />.
-///     Обозначает некоторый доступ к некоторой функции.
+///     Роль для <see cref="User" />.
 /// </summary>
-public class Permission : Identity
+public class Role : Identity
 {
     /*                   __ _                       _   _
      *   ___ ___  _ __  / _(_) __ _ _   _ _ __ __ _| |_(_) ___  _ __
@@ -33,9 +31,9 @@ public class Permission : Identity
     public const bool IsNameRequired = true;
 
     /// <summary>
-    ///     Конфигурация модели <see cref="Permission" />.
+    ///     Конфигурация модели <see cref="User" />.
     /// </summary>
-    internal class Configuration : Configuration<Permission>
+    internal class Configuration : Configuration<Role>
     {
         /// <summary>
         ///     Конструктор.
@@ -50,15 +48,19 @@ public class Permission : Identity
         ///     Задать конфигурацию для модели.
         /// </summary>
         /// <param name="builder">Набор интерфейсов настройки модели.</param>
-        public override void Configure(EntityTypeBuilder<Permission> builder)
+        public override void Configure(EntityTypeBuilder<Role> builder)
         {
-            builder.Property(permission => permission.Name)
-                .HasMaxLength(NameLengthMax)
+            builder.Property(role => role.Name)
+                .HasMaxLength(User.NameLengthMax)
                 .IsRequired();
 
-            builder.HasMany(permission => permission.RolePermissionLinks)
-                .WithOne(link => link.Permission)
-                .HasForeignKey(link => link.PermissionId);
+            builder.HasMany(role => role.Users)
+                .WithOne(user => user.Role)
+                .HasForeignKey(user => user.RoleId);
+
+            builder.HasMany(role => role.RolePermissionLinks)
+                .WithOne(link => link.Role)
+                .HasForeignKey(link => link.RoleId);
 
             base.Configure(builder);
         }
@@ -84,6 +86,11 @@ public class Permission : Identity
     public string Name { get; set; } = null!;
 
     #endregion
+
+    /// <summary>
+    ///     Связанные объекты <see cref="User" />.
+    /// </summary>
+    public List<User> Users { get; set; } = null!;
 
     /// <summary>
     ///     Связанные объекты <see cref="RolePermissionLink" />.
