+++ Studenda.Core/Model/Security/User.cs
@@ -1,8 +1,9 @@
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Studenda.Core.Data.Configuration;
 using Studenda.Core.Model.Link;
+using Studenda.Core.Model.Security.Management;
 
-namespace Studenda.Core.Model.Account;
+namespace Studenda.Core.Model.Security;
 
 /// <summary>
 ///     Пользователь.
